pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.33.0",
    "gunicorn>=23.0.0",
    "langdetect>=1.0.9",
    "openai>=1.106.1",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
